// src/app/features/orders/models/order.enums.ts
export enum OrderStatus {
  RECEIVED = 'Received',
  PREPARING = 'Preparing',
  READY = 'Ready',
  DELIVERED = 'Delivered',
  COMPLETED = 'Completed'
}

export enum OrderChannel {
  WALK_IN = 'Walk-in',
  DELIVERY = 'Delivery',
  ONLINE = 'Online'
}

export enum OrderPriority {
  LOW = 'Low',
  MEDIUM = 'Medium',
  HIGH = 'High'
}
export enum Priority {
  LOW = 'Low',
  MEDIUM = 'Medium',
  HIGH = 'High'
}